// src/types/app.ts
export enum TransferMethod {
    all = 'all',
    local_file = 'local_file',
    remote_url = 'remote_url'
}

export type VisionSettings = {
    enabled: boolean;
    number_limits: number;
    detail: Resolution;
    transfer_methods: TransferMethod[];
    image_file_size_limit?: number | string;
};

export type ImageFile = {
    type: TransferMethod;
    _id: string;
    fileId: string;
    file?: File;
    progress: number;
    url: string;
    base64Url?: string;
    deleted?: boolean;
};